requirements_tumor_classifier.txt
streamlit>=1.22.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.26.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.64.0
scikit-learn>=1.0.0 